pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comet-api"
version = "0.1.0"
description = "Multi-region ride-hailing platform — API foundation (modular monolith)"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.1",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "psycopg[binary]==3.2.3",
    "alembic==1.14.1",
    "redis==5.2.1",
    "structlog==24.4.0",
    "newrelic==10.3.1",
    "PyJWT[crypto]==2.10.1",
    "slowapi==0.1.9",
    "celery[redis]==5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.0.0",
    "httpx==0.28.1",
    "ruff==0.14.14",
    "pre-commit==4.0.1",
    "locust==2.32.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]
"app/api/v1/endpoints/*.py" = ["B008"]
"tests/conftest.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
